+++ commands/root.go
@@ -11,7 +11,7 @@ import (
 	"github.com/ionos-cloud/ionosctl/v6/commands/cdn"
 	"github.com/ionos-cloud/ionosctl/v6/commands/kafka"
 
-	certificates "github.com/ionos-cloud/ionosctl/v6/commands/certmanager"
+	certificates "github.com/ionos-cloud/ionosctl/v6/commands/cert"
 	"github.com/ionos-cloud/ionosctl/v6/commands/cfg"
 	cloudapiv6 "github.com/ionos-cloud/ionosctl/v6/commands/cloudapi-v6"
 	container_registry "github.com/ionos-cloud/ionosctl/v6/commands/container-registry"
@@ -219,7 +219,7 @@ func addCommands() {
 	// Add DBaaS Commands
 	rootCmd.AddCommand(dbaas.DataBaseServiceCmd())
 	// Add Certificate Manager Commands
-	rootCmd.AddCommand(certificates.CertCmd())
+	rootCmd.AddCommand(certificates.Root())
 	// Dataplatform commands
 	rootCmd.AddCommand(dataplatform.DataplatformCmd())
 	// Add Container Registry Commands